Preparing to release version 0.1.2.

+++ library/kmplitert-core/src/jsMain/kotlin/io/github/leitingzi/kmplitert/core/LiteRTCompiler.js.kt
@@ -153,7 +153,9 @@ actual class LiteRTCompiler actual constructor(
     }
 
     actual suspend fun close() {
-        compiledModel.delete()
+        if (::compiledModel.isInitialized) {
+            compiledModel.delete()
+        }
     }
 
     private fun getCompileOptions(accelerator: LiteRTAccelerator): JsAny {
