+++ library/kmplitert-core/src/wasmJsMain/kotlin/io/github/leitingzi/kmplitert/core/LiteRTCompiler.wasmJs.kt
@@ -160,7 +160,9 @@ actual class LiteRTCompiler actual constructor(
     }
 
     actual suspend fun close() {
-        compiledModel.delete()
+        if (::compiledModel.isInitialized) {
+            compiledModel.delete()
+        }
     }
 
     private fun getCompileOptions(accelerator: LiteRTAccelerator): JsAny {
